feat: PHPStan 0 erreurs + E2E CI
- PHPStan baseline 132→0 erreurs (level 8)
- 21 erreurs PDOStatement|false corrigées (migrations v13-v26)
- 32 erreurs type safety corrigées (20 fichiers src/)
- Tests E2E ajoutés au pipeline Woodpecker
- Tests: 977→1292, assertions: 1628→2119

+++ src/Render/HtmlService.php
@@ -139,7 +139,7 @@ final class HtmlService implements HtmlInterface
 
         // Cas 2 : masquer le domaine si = domaine du user courant
         if ($current_user !== '' && str_contains($current_user, '@')) {
-            $user_domain = strtolower(substr($current_user, strrpos($current_user, '@')));
+            $user_domain = strtolower(substr($current_user, (int) strrpos($current_user, '@')));
             $email_lower = strtolower($email);
             if (str_ends_with($email_lower, $user_domain)) {
                 $local_part = substr($email, 0, strlen($email) - strlen($user_domain));
